Create a simple UI.
Navigating to / will display a table of all registered go-links, and /<foo> where foo is undefined will go to a form to register the link.
The next step is to wire things up so that it actually adds a new link.
Future work might be to add an "edit" button (hence the inclusion of HTMX) and to style the page.